Update to allow dump/apply of only one object type.
- add --object_type option to specify which object type to dump/apply.

+++ utils/make_dir.js
@@ -5,12 +5,15 @@ const path = require('path');
 const debug = require('debug')('kcm:make_dir');
 const logger = require('./logger');
 
-module.exports = function makeDir(dirName) {
+module.exports = function makeDir(dirName, remove = true) {
   debug(`Ready to make dir ${dirName}...`);
   const dirPath = path.resolve(process.cwd(), dirName);
   debug(`Dir path is ${dirPath}`);
   try {
-    fse.removeSync(dirPath);
+    if (remove) {
+      debug(`Removing dir ${dirName}...`);
+      fse.removeSync(dirPath);
+    }
     fse.ensureDirSync(dirPath);
   } catch (e) {
     logger.error(`Fail to make dir ${dirName}: ${e}`);
